add junit test

+++ src/controller/api.ts
@@ -14,57 +14,50 @@ import {
   Validate,
  } from '@midwayjs/decorator';
 import { Context } from 'egg';
-import { IGetUserResponse } from '../interface';
 import { IGetTodoItemResponse } from '../interface';
 import { TodoItemService } from '../service/todoItem';
 import { TNgaTodoItem } from '../entity/TNgaTodoItem';
-import { UserService } from '../service/user';
 
 
 @Provide()
-@Controller('/api', { middleware: [ 'reportMiddleware' ] })
+@Controller('/api/todo', { middleware: [ 'reportMiddleware' ] })
 export class APIController {
   @Inject()
   ctx: Context;
 
   @Inject()
-  userService: UserService;
-
-  @Inject()
   todoItemService: TodoItemService;
 
-  @Post('/user/get_user', { middleware: [ 'reportMiddleware' ] })
-  async getUser(@Query() uid: string): Promise<IGetUserResponse> {
-    const user = await this.userService.getUser({ uid });
-    console.log("------getUser------");
-    return { success: true, message: 'OK', data: user };
-  }
-
-  @Get('/todo/get')
+  @Get('/get')
   async get(@Query() id: number): Promise<IGetTodoItemResponse> {
+    if (id == null)
+      throw new Error('id is null');
     const todoItems = await this.todoItemService.get(id);
     return { success: true, message: 'OK', data: todoItems };
   }
 
-  @Post('/todo/save')
+  @Post('/save')
   @Validate()
   async save(@Body(ALL) todoItem: TNgaTodoItem): Promise<IGetTodoItemResponse> {
     const result = await this.todoItemService.save(todoItem);
     return { success: true, message: 'OK', data: result };
   }
 
-  @Put('/todo/put')
+  @Put('/put')
   async put(@Body(ALL) todoItem: TNgaTodoItem): Promise<IGetTodoItemResponse> {
     const result = await this.todoItemService.put(todoItem);
     return { success: true, message: 'OK', data: result };
   }
-  @Get('/todo/list')
+  @Get('/list')
   async list(): Promise<IGetTodoItemResponse> {
     const todoItems = await this.todoItemService.list();
     return { success: true, message: 'OK', data: todoItems };
   }
-  @Del('/todo/delete')
+
+  @Del('/delete')
   async delete(@Query() id: number): Promise<IGetTodoItemResponse> {
+    if (id == null )
+      throw new Error('id is null');
     const todoItems = await this.todoItemService.delete(id);
     return { success: true, message: 'OK', data: todoItems };
   }
